% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cyclone_energy.R
\name{cyclone_energy}
\alias{cyclone_energy}
\title{Computing the accumulated cyclone energy}
\usage{
cyclone_energy(storm_id)
}
\arguments{
\item{storm_id}{object of type character containing a storm id}
}
\value{
a scalar representing the accumulated energy of the storm with the
requested storm id
}
\description{
Given an id of a storm, return the accumulated energy, which is the sum of
the squared maximum sustained velocities recorded every 6 hours.
}
\examples{
AL141933.energy <- cyclone_energy("AL141933")
print(AL141933.energy)
}
