% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/interpolate.R
\name{interpolate}
\alias{interpolate}
\title{Storm Track Interpolation}
\usage{
interpolate(storm_id)
}
\arguments{
\item{storm_id}{object of type character containing a storm id}
}
\value{
a data frame containing the predicted longitude and latitude
coordinates every 30 minutes derived using linear interpolation
}
\description{
Given the id of a storm, predict its longitude and latitude coordinates every
30 minutes using linear interpolation.
}
\examples{
interp_storm <- interpolate("AL252005")
print(interp_storm)
}
