% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/made_landfall.R
\name{made_landfall}
\alias{made_landfall}
\title{Determine whether a storm made landfall}
\usage{
made_landfall(storm_id, time.pt = NULL)
}
\arguments{
\item{storm_id}{object of type character containing a storm id}

\item{time.pt}{object of type POSIXct indicating a particular time point
of interest in the storm's trajectory at which to test whether the storm
made landfall; NULL by default}
}
\value{
an object of type logical indicating whether the storm made landfall,
i.e., whether the storm reached the United States, including non-contiguous states
and surrounding islands
}
\description{
Given the id of a storm, return whether the storm made landfall.
}
\examples{
library(maps); library(sp)
landfall.logical <- made_landfall("AL032019", as.POSIXct("2019-07-22 12:00:00", tz='UTC'))
print(landfall.logical)
}
