% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/storm_location.R
\name{storm_location}
\alias{storm_location}
\title{Storm Track Size and Location}
\usage{
storm_location(storm_id, show.track = FALSE, time.pt = NULL)
}
\arguments{
\item{storm_id}{object of type character containing a storm id}

\item{show.track}{a logical argument indicating whether the track of the storm
should be plotted on the map; FALSE by default}

\item{time.pt}{object of type POSIXct indicating a particular time point
of interest in the storm's trajectory that the map will plot; NULL by default,
and in this case, plots of storm's size and shape will be overlaid on the map
at every time point}
}
\value{
a map plotting the trajectory and size of the storm, as indicated by
regions for which the maximum wind extent is at least 34, 50, and 64 knots
}
\description{
Given the id of a storm, return a map of its trajectory and size over time.
}
\examples{
library(maps)
storm.loc.ex <- storm_location("AL182012")

}
