% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/storm_track.R
\name{storm_track}
\alias{storm_track}
\title{Plotting the Storm Track}
\usage{
storm_track(storm_vec, show.legend = FALSE)
}
\arguments{
\item{storm_vec}{vector of type character containing storm ids}

\item{show.legend}{logical argument indicating whether a legend with the storm ids
should be provided; if the length of storm_vec is less than 5, it is plotted on the
map, and if it is greater than 5, it is returned separately as a data frame}
}
\value{
a color-coded map of the trajectories, or tracks, of all of the storms in
storm_vec using the available data
}
\description{
Given a vector of storm ids, return a map of the trajectories, or tracks, of
all of the storms using the available data.
}
\examples{
library(maps)
ex.ids <- c("AL042005", "AL051897", "AL162016")
ex.tracks <- storm_track(ex.ids, TRUE)
}
